requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
dj-database-url==2.1.0
Django==5.0.6
django-allauth==0.63.3
django-colorfield==0.11.0
django-crispy-forms==2.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
gunicorn==22.0.0
oauthlib==3.2.2
pillow==10.3.0
psycopg2==2.9.6
PyJWT==2.8.0
python-slugify==8.0.4
python3-openid==3.2.0
requests-oauthlib==2.0.0
sqlparse==0.5.0
text-unidecode==1.3
whitenoise==6.7.0
